+++ src/main/java/com/sevensenders/cpt/codingchallenge/model/Order.java
@@ -1,31 +1,23 @@
 package com.sevensenders.cpt.codingchallenge.model;
 
+
 import lombok.Data;
 
-
-import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+
+import javax.persistence.Entity;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.Instant;
 
 @Data
-@Entity
-public class Shipment implements Serializable {
+@Entity(name = "OrderEntity")
+public class Order implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer shipmentId;
-
-    private String trackingCode;
-
-    private String warehouseAddress;
-
-    private String shipmentStatus;
-
-    private BigDecimal weight;
+    private Integer orderId;
 
     private Instant createdAt;
